Multiplication Table

+++ 219432_F.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char const *argv[])
 {
-    uint8_t x;
+    short x;
     std::cin >> x;
     for (size_t i = 1; i <= 12; i++)
         std::cout << x << " * " << i << " = " << x * i << "\n";
